Make TextSpan u32 to match the compiler one (#1748)
commit-id:107b5e17

---

**Stack**:
- #1749
- #1748 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*

+++ scarb/tests/build_cairo_plugin.rs
@@ -414,7 +414,7 @@ fn can_replace_original_node() {
             let new_token_string = token_stream.to_string().replace("12", "34");
             let token_stream = TokenStream::new(vec![TokenTree::Ident(Token::new(
                 new_token_string.clone(),
-                TextSpan { start: 0, end: new_token_string.len() },
+                TextSpan { start: 0, end: new_token_string.len() as u32 },
             ))]);
             ProcMacroResult::new(token_stream)
         }
@@ -585,7 +585,7 @@ fn can_define_multiple_macros() {
             let new_token_string = token_stream.to_string().replace("12", "34");
             let token_stream = TokenStream::new(vec![TokenTree::Ident(Token::new(
                 new_token_string.clone(),
-                TextSpan { start: 0, end: new_token_string.len() },
+                TextSpan { start: 0, end: new_token_string.len() as u32 },
             ))]);
             let aux_data = AuxData::new(Vec::new());
             ProcMacroResult::new(token_stream).with_aux_data(aux_data)
@@ -596,7 +596,7 @@ fn can_define_multiple_macros() {
             let new_token_string = token_stream.to_string().replace("56", "78");
             let token_stream = TokenStream::new(vec![TokenTree::Ident(Token::new(
                 new_token_string.clone(),
-                TextSpan { start: 0, end: new_token_string.len() },
+                TextSpan { start: 0, end: new_token_string.len() as u32 },
             ))]);
             let aux_data = AuxData::new(Vec::new());
             ProcMacroResult::new(token_stream).with_aux_data(aux_data)
@@ -620,7 +620,7 @@ fn can_define_multiple_macros() {
             let new_token_string = token_stream.to_string().replace("90", "09");
             let token_stream = TokenStream::new(vec![TokenTree::Ident(Token::new(
                 new_token_string.clone(),
-                TextSpan { start: 0, end: new_token_string.len() },
+                TextSpan { start: 0, end: new_token_string.len() as u32 },
             ))]);
             let aux_data = AuxData::new(Vec::new());
             ProcMacroResult::new(token_stream).with_aux_data(aux_data)
@@ -825,7 +825,7 @@ fn can_resolve_full_path_markers() {
               code.clone(),
                 TextSpan {
                   start: 0,
-                  end: code.len(),
+                  end: code.len() as u32,
                 },
               ))])
             ).with_full_path_markers(full_path_markers)
@@ -998,7 +998,7 @@ fn can_implement_derive_macro() {
                   code.clone(),
                     TextSpan {
                         start: 0,
-                        end: code.len(),
+                        end: code.len() as u32,
                     },
                 ))]);
 
@@ -1061,7 +1061,7 @@ fn can_use_both_derive_and_attr() {
                   new_token_string.clone(),
                     TextSpan {
                         start: 0,
-                        end: new_token_string.len(),
+                        end: new_token_string.len() as u32,
                     },
                 ))]))
             }
@@ -1073,7 +1073,7 @@ fn can_use_both_derive_and_attr() {
                   code.clone(),
                     TextSpan {
                         start: 0,
-                        end: code.len(),
+                        end: code.len() as u32,
                     },
                 ))]);
 
@@ -1082,7 +1082,7 @@ fn can_use_both_derive_and_attr() {
                   result_string.clone(),
                     TextSpan {
                         start: 0,
-                        end: result_string.len(),
+                        end: result_string.len() as u32,
                     },
                 ))]))
             }
@@ -1101,7 +1101,7 @@ fn can_use_both_derive_and_attr() {
                   code.clone(),
                     TextSpan {
                         start: 0,
-                        end: code.len(),
+                        end: code.len() as u32,
                     },
                 ))]))
             }
@@ -1310,7 +1310,7 @@ fn can_be_expanded() {
             let new_token_string = token_stream.to_string().replace("12", "34");
             let token_stream = TokenStream::new(vec![TokenTree::Ident(Token::new(
                 new_token_string.clone(),
-                TextSpan { start: 0, end: new_token_string.len() },
+                TextSpan { start: 0, end: new_token_string.len() as u32 },
             ))]);
             ProcMacroResult::new(token_stream)
         }
@@ -1340,7 +1340,7 @@ fn can_be_expanded() {
 
             let token_stream = TokenStream::new(vec![TokenTree::Ident(Token::new(
                 code.clone(),
-                TextSpan { start: 0, end: code.len() },
+                TextSpan { start: 0, end: code.len() as u32 },
             ))]);
 
             ProcMacroResult::new(token_stream)
@@ -1426,7 +1426,7 @@ fn can_expand_trait_inner_func_attrr() {
                     .replace("12", "34");
                 ProcMacroResult::new(TokenStream::new(vec![TokenTree::Ident(Token::new(
                   new_token_string.clone(),
-                  TextSpan { start: 0, end: new_token_string.len() },
+                  TextSpan { start: 0, end: new_token_string.len() as u32 },
                 ))]))
             }
         "##})
@@ -1488,7 +1488,7 @@ fn can_expand_impl_inner_func_attrr() {
                 let new_token_string = token_stream.to_string().replace("1", "2");
                 ProcMacroResult::new(TokenStream::new(vec![TokenTree::Ident(Token::new(
                     new_token_string.clone(),
-                    TextSpan { start: 0, end: new_token_string.len() },
+                    TextSpan { start: 0, end: new_token_string.len() as u32 },
                 ))]))
             }
         "##})
